+++ serverfolder/POS_Server/Controllers/AgentMembershipCashController.cs
@@ -376,13 +376,13 @@ namespace POS_Server.Controllers
             }
             else
             {
-                int agentMembershipCashId = 0;
+                long agentMembershipCashId = 0;
                 IEnumerable<Claim> claims = TokenManager.getTokenClaims(token);
                 foreach (Claim c in claims)
                 {
                     if (c.Type == "itemId")
                     {
-                        agentMembershipCashId = int.Parse(c.Value);
+                        agentMembershipCashId = long.Parse(c.Value);
                     }
                 }
                 using (incposdbEntities entity = new incposdbEntities())
@@ -444,32 +444,32 @@ namespace POS_Server.Controllers
                 }
                 if (newObject.updateUserId == 0 || newObject.updateUserId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.updateUserId = id;
                 }
                 if (newObject.createUserId == 0 || newObject.createUserId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.createUserId = id;
                 }
                 if (newObject.subscriptionFeesId == 0 || newObject.subscriptionFeesId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.subscriptionFeesId = id;
                 }
                 if (newObject.cashTransId == 0 || newObject.cashTransId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.cashTransId = id;
                 }
                 if (newObject.membershipId == 0 || newObject.membershipId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.membershipId = id;
                 }
                 if (newObject.agentId == 0 || newObject.agentId == null)
                 {
-                    Nullable<int> id = null;
+                    Nullable<long> id = null;
                     newObject.agentId = id;
                 }
                 try
@@ -539,19 +539,19 @@ namespace POS_Server.Controllers
             }
             else
             {
-                int agentMembershipCashId = 0;
-                int userId = 0;
+                long agentMembershipCashId = 0;
+                long userId = 0;
                 Boolean final = false;
                 IEnumerable<Claim> claims = TokenManager.getTokenClaims(token);
                 foreach (Claim c in claims)
                 {
                     if (c.Type == "itemId")
                     {
-                        agentMembershipCashId = int.Parse(c.Value);
+                        agentMembershipCashId = long.Parse(c.Value);
                     }
                     else if (c.Type == "userId")
                     {
-                        userId = int.Parse(c.Value);
+                        userId = long.Parse(c.Value);
                     }
                     else if (c.Type == "final")
                     {
@@ -614,8 +614,8 @@ namespace POS_Server.Controllers
             {
                 string strObject = "";
                 List<agentMembershipCash> newListObj = null;
-                int membershipId = 0;
-                int updateUserId = 0;
+                long membershipId = 0;
+                long updateUserId = 0;
                 IEnumerable<Claim> claims = TokenManager.getTokenClaims(token);
                 foreach (Claim c in claims)
                 {
@@ -628,12 +628,12 @@ namespace POS_Server.Controllers
                     }
                     else if (c.Type == "membershipId")
                     {
-                        membershipId = int.Parse(c.Value);
+                        membershipId = long.Parse(c.Value);
                     }
                     else
                   if (c.Type == "updateUserId")
                     {
-                        updateUserId = int.Parse(c.Value);
+                        updateUserId = long.Parse(c.Value);
                     }
                 }
 
@@ -664,32 +664,32 @@ namespace POS_Server.Controllers
                         {
                             if (newListObj[i].updateUserId == 0 || newListObj[i].updateUserId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].updateUserId = id;
                             }
                             if (newListObj[i].createUserId == 0 || newListObj[i].createUserId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].createUserId = id;
                             }
                             if (newListObj[i].membershipId == 0 || newListObj[i].membershipId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].membershipId = id;
                             }
                             if (newListObj[i].agentId == 0 || newListObj[i].agentId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].agentId = id;
                             }
                             if (newListObj[i].subscriptionFeesId == 0 || newListObj[i].subscriptionFeesId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].subscriptionFeesId = id;
                             }
                             if (newListObj[i].cashTransId == 0 || newListObj[i].cashTransId == null)
                             {
-                                Nullable<int> id = null;
+                                Nullable<long> id = null;
                                 newListObj[i].cashTransId = id;
                             }
                             var branchEntity = entity.Set<agentMembershipCash>();
@@ -736,8 +736,8 @@ namespace POS_Server.Controllers
         //    {
         //        string strObject = "";
         //        List<agentMembershipCash> newListObj = null;
-        //        int membershipId = 0;
-        //        int updateUserId = 0;
+        //        long membershipId = 0;
+        //        long updateUserId = 0;
         //        IEnumerable<Claim> claims = TokenManager.getTokenClaims(token);
         //        foreach (Claim c in claims)
         //        {
@@ -750,12 +750,12 @@ namespace POS_Server.Controllers
         //            }
         //            else if (c.Type == "membershipId")
         //            {
-        //                membershipId = int.Parse(c.Value);
+        //                membershipId = long.Parse(c.Value);
         //            }
         //            else
         //          if (c.Type == "updateUserId")
         //            {
-        //                updateUserId = int.Parse(c.Value);
+        //                updateUserId = long.Parse(c.Value);
         //            }
         //        }
 
@@ -834,32 +834,32 @@ namespace POS_Server.Controllers
         //    int message = 0;
         //        if (newObject.updateUserId == 0 || newObject.updateUserId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.updateUserId = id;
         //        }
         //        if (newObject.createUserId == 0 || newObject.createUserId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.createUserId = id;
         //        }
         //        if (newObject.subscriptionFeesId == 0 || newObject.subscriptionFeesId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.subscriptionFeesId = id;
         //        }
         //        if (newObject.cashTransId == 0 || newObject.cashTransId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.cashTransId = id;
         //        }
         //        if (newObject.membershipId == 0 || newObject.membershipId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.membershipId = id;
         //        }
         //        if (newObject.agentId == 0 || newObject.agentId == null)
         //        {
-        //            Nullable<int> id = null;
+        //            Nullable<long> id = null;
         //            newObject.agentId = id;
         //        }
         //        try
@@ -921,37 +921,37 @@ namespace POS_Server.Controllers
 
 
 
-        public int Saveamc(agentMembershipCash newObject)
+        public long Saveamc(agentMembershipCash newObject)
         {
-            int message = 0;
+            long message = 0;
             if (newObject.updateUserId == 0 || newObject.updateUserId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.updateUserId = id;
             }
             if (newObject.createUserId == 0 || newObject.createUserId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.createUserId = id;
             }
             if (newObject.subscriptionFeesId == 0 || newObject.subscriptionFeesId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.subscriptionFeesId = id;
             }
             if (newObject.cashTransId == 0 || newObject.cashTransId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.cashTransId = id;
             }
             if (newObject.membershipId == 0 || newObject.membershipId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.membershipId = id;
             }
             if (newObject.agentId == 0 || newObject.agentId == null)
             {
-                Nullable<int> id = null;
+                Nullable<long> id = null;
                 newObject.agentId = id;
             }
             try
@@ -1045,7 +1045,7 @@ namespace POS_Server.Controllers
 
                 try
                 {
-                    int cashTransId = 0;
+                    long cashTransId = 0;
                     DateTime lastenddate = DateTime.Now;
                     DateTime DTNow = DateTime.Now;
                     DateTime tmp;
@@ -1057,7 +1057,7 @@ namespace POS_Server.Controllers
                     subscriptionFeesController subctrlr = new subscriptionFeesController();
                     // get month count
 
-                    subscriptionmodel = subctrlr.GetById((int)newObject.subscriptionFeesId);
+                    subscriptionmodel = subctrlr.GetById((long)newObject.subscriptionFeesId);
                     // return TokenManager.GenerateToken(subscriptionmodel);
                     newObject.monthsCount = subscriptionmodel.monthsCount;
 
@@ -1187,7 +1187,7 @@ namespace POS_Server.Controllers
 
                         newObject.cashTransId = cashTransId;
                         //save AgentMembershipCash 
-                        int res = Saveamc(newObject);
+                        long res = Saveamc(newObject);
                         return TokenManager.GenerateToken(res.ToString());
                     }
                     else
